+++ src/MessageBox.Avalonia/MessageBoxManager.cs
@@ -1,15 +1,12 @@
 using Avalonia.Controls;
 using MessageBox.Avalonia.BaseWindows;
+using MessageBox.Avalonia.BaseWindows.Base;
 using MessageBox.Avalonia.DTO;
 using MessageBox.Avalonia.Enums;
 using MessageBox.Avalonia.ViewModels;
 using CustomWindow = MessageBox.Avalonia.Views.MsBoxCustomWindow;
 using HyperlinkWindow = MessageBox.Avalonia.Views.MsBoxHyperlinkWindow;
 using InputWindow = MessageBox.Avalonia.Views.MsBoxInputWindow;
-using MsBoxCustomWindow = MessageBox.Avalonia.BaseWindows.MsBoxCustomWindow;
-using MsBoxHyperlinkWindow = MessageBox.Avalonia.BaseWindows.MsBoxHyperlinkWindow;
-using MsBoxInputWindow = MessageBox.Avalonia.BaseWindows.MsBoxInputWindow;
-using MsBoxStandardWindow = MessageBox.Avalonia.BaseWindows.MsBoxStandardWindow;
 using StandardWindow = MessageBox.Avalonia.Views.MsBoxStandardWindow;
 
 namespace MessageBox.Avalonia
@@ -21,7 +18,7 @@ namespace MessageBox.Avalonia
             var window = new CustomWindow(@params.Style);
             @params.Window = window;
             window.DataContext = new MsBoxCustomViewModel(@params);
-            return new MsBoxCustomWindow(window);
+            return new MsBoxWindowBase<CustomWindow, string>(window, window.ButtonResult);
         }
 
         public static IMsBoxWindow<ButtonResult> GetMessageBoxStandardWindow(MessageBoxStandardParams @params)
@@ -29,13 +26,13 @@ namespace MessageBox.Avalonia
             var window = new StandardWindow(@params.Style);
             @params.Window = window;
             window.DataContext = new MsBoxStandardViewModel(@params);
-            return new MsBoxStandardWindow(window);
+            return new MsBoxWindowBase<StandardWindow,ButtonResult>(window, window.ButtonResult);
         }
         public static IMsBoxWindow<ButtonResult> GetMessageBoxHyperlinkWindow(MessageBoxHyperlinkParams @params)
         {
             var window = new HyperlinkWindow(@params.Style) { DataContext = new MsBoxHyperlinkViewModel(@params) };
             @params.Window = window;
-            return new MsBoxHyperlinkWindow(window);
+            return new  MsBoxWindowBase<HyperlinkWindow, ButtonResult>(window, window.ButtonResult);
         }
         public static IMsBoxWindow<ButtonResult> GetMessageBoxStandardWindow(string title, string text,
             ButtonEnum @enum = ButtonEnum.Ok, Icon icon = Icon.None,
@@ -55,7 +52,7 @@ namespace MessageBox.Avalonia
             var window = new InputWindow(@params.Style);
             @params.Window = window;
             window.DataContext = new MsBoxInputViewModel(@params);
-            return new MsBoxInputWindow(window);
+            return new MsBoxWindowBase<InputWindow, MessageWindowResultDTO>(window, new MessageWindowResultDTO(window.MessageResult, window.ButtonResult));
         }
     }
 }
